+++ jenkviz/util.py
@@ -82,7 +82,7 @@ def duration_to_second(duration):
     """Convert jenkins duration into second"""
     if not duration:
         return None
-    match = re.match('^(([0-9])+ hr)? ?(([0-9]+) min)? ?(([0-9]+) sec)?$', duration)
+    match = re.match('^(([0-9])+ hr)? ?(([0-9]+) min)? ?(([0-9\.]+) sec)?$', duration)
     ret = 0
     if match and len(match.groups()) == 6:
         if match.group(2):
@@ -90,7 +90,7 @@ def duration_to_second(duration):
         if match.group(4):
             ret += 60 * int(match.group(4))
         if match.group(6):
-            ret += int(match.group(6))
+            ret += int(float(match.group(6)))
     return ret