+++ test.py
@@ -17,6 +17,8 @@ from jenkviz.main import main
 
 class JenkVizTestCase(TestCase):
     def test_duration(self):
+        self.assertEquals(duration_to_second('8 sec'), 8)
+        self.assertEquals(duration_to_second('8.8 sec'), 8)
         self.assertEquals(duration_to_second('1 min'), 60)
         self.assertEquals(duration_to_second('1 min 1 sec'), 61)
         self.assertEquals(duration_to_second('10 min 10 sec'), 610)